backend/pyproject.toml
[project]
name = "ai-stock-tracker-backend"
version = "0.1.0"
description = "AI Stock Tracker Backend with Foundation Model Forecasting"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "yfinance>=0.2.50",
    "sqlalchemy>=2.0.36",
    "pydantic>=2.10.0",
    "torch>=2.5.0",
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.6.0",
    "transformers>=4.47.0",
    "accelerate>=1.2.0",
    "timesfm @ git+https://github.com/google-research/timesfm.git",
    "chronos-forecasting @ git+https://github.com/amazon-science/chronos-forecasting.git",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.28.0",
]

[tool.uv]
package = false

